Made the shop class. Alkso changed fame to double instead of int, resulting in bugs that will be solved in the future

+++ Game/Game/Character.cpp
@@ -31,7 +31,7 @@ string Character::getName()
 }
 
 //Getter for fame, returns fame
-int Character::getFame()
+double Character::getFame()
 {
     return fame;
 }
@@ -73,7 +73,7 @@ void Character::setName(string nomen)
 }
 
 //Setter for damage, sets fame equal to influence
-void Character::setFame(int influence)
+void Character::setFame(double influence)
 {
     fame = influence;
 }
