+++ Game/Game/Enemy.h
@@ -6,7 +6,7 @@ protected:
     //will get more as time progresses
 public:
     //The same constructor, for now, as the character class
-    Enemy(int life, int attack, int def, string nomen, int influence, int anno) :Character(life, attack, def, nomen, influence, anno) {};
+    Enemy(int life, int attack, int def, string nomen, double influence, int anno) :Character(life, attack, def, nomen, influence, anno) {};
 
     static Enemy generateEnemy(); //generates random enemy.  takes nothing in, creates an enemy with randiomly selected name, randomizes other character stats
     Attack chooseAttack() override; //function that overides virtual function in character
